Commited week 2 Programs
For the following task given, you should upload your program on your Github account. If your upload is in between 9:00am to 10:50pm on Saturday, 24-8-2019 then your score will be 10marks. Else if your upload is on or before 30-8-2019 then your score will be 6marks. Submissions after 30-8-2019 will be awarded with Zero marks.

Programs based on Selection Sort and Bubble Sort
2a.
Write a program for printing kth smallest elements in an array,
Given an array and a number k where k is smaller than size of array, we need to find the kth smallest element in the given array. It is given that all array elements are distinct.
Examples:
Input: arr[] = {7, 10, 4, 3, 20, 15} k = 3 Output: 7
Input: arr[] = {7, 10, 4, 3, 20, 15} k = 4 Output: 10

2b.
Write an efficient program for printing k largest elements in an array. Elements in the array can be in any order. It is given that all array elements are distinct.
For example, if given array is [1, 23, 12, 9, 30, 2, 50] and you are asked for the largest 3 elements i.e., k = 3 then your program should print 50, 30 and 23.